+++ R/validation-uniqueness.R
@@ -1,4 +1,4 @@
-#' Check that an array of values does not include any duplicate
+#' Check that a vector has only unique values
 #'
 #' @param values An array of values
 #' @param na.rm Logical to decide if missing values should be removed from the
@@ -8,6 +8,7 @@
 #' @return Logical assessing if all values are unique.
 #'
 #' @examples
+#'
 #' hasOnlyDistinctValues(c("x", "y"))
 #' hasOnlyDistinctValues(c("x", "y", "x"))
 #' hasOnlyDistinctValues(c("x", NA, "y", NA), na.rm = FALSE)
@@ -20,4 +21,23 @@ hasOnlyDistinctValues <- function(values, na.rm = TRUE) {
   }
 
   return(!any(duplicated(values)))
+}
+
+#' Validate that vector has unique values
+#'
+#' @inheritParams hasOnlyDistinctValues
+#'
+#' @examples
+#' validateHasOnlyDistinctValues(c("x", "y")) # NULL
+#' # validateHasOnlyDistinctValues(c("x", "y", "x")) # error
+#'
+#' @return `NULL` if only unique values present, otherwise produces error.
+#'
+#' @export
+validateHasOnlyDistinctValues <- function(values, na.rm = TRUE) {
+  if (hasOnlyDistinctValues(values)) {
+    return()
+  }
+
+  stop(messages$errorDuplicatedValues())
 }