+++ NEWS.md
@@ -6,6 +6,8 @@ NEW FUNCTIONS
 
 * `objectCount` to count number of objects.
 
+* `validateHasOnlyDistinctValues()` to validate only unique values are present.
+
 MAJOR CHANGES
 
 * Cloning method is now disabled for `Printable` R6 class. This entails that